requirements.txt
ldap3>=2.9
PyYAML>=6.0
typer[all]>=0.9
